Add initial Media client
Adds functions to retrieve a page of Media resources, to get a individual Media
instance, and to get a URL where Media can be downloaded. Note that Twilio
gives us back an insecure HTTP URL for Media. We have to rewrite it from the
virtual-host based S3 format to the path-based format in order to retrieve the
resource securely.