Use the same jwt library everywhere
Upgrade the existing twilioclient library to the one used in the token package.
Fixes an error in token (and in twilioclient) where the "Issued" timestamp
could be ahead or behind of the "Expires" timestamp based on a timezone
difference.

I'm still not happy that the API's here differ, I would like to make them
more consistent, at least in terms of how the JWT token is generated, but at
least there's only one JWT library and not two.

Fixes #2.