Don't pass range filters if they're not needed
Too early date range or too late of a date range means that we might trigger an
unnecessary WHERE clause in mysql on the backend. Instead add another sentinel
value, so users can indicate when they don't want a query to be filtered on one
side, and omit filters when they provide it.

Rewrite every use of GetPage to call GetPageIterator(data).Next(ctx)
internally, there's no need for a separate implementation. Fixes #9.