Use the new rest.DefaultTransport
This makes it easier to debug HTTP requests and allows for a little more
flexibility about how they are debugged and where the output goes.

Also fix the NewMonitorClient and NewPricingClient to use the provided
http.Client objects instead of just ignoring them, and updates the build matrix
to Go 1.7 only.