Implement v2 pricing API
The API update requires several breaking changes, which are documented
in the CHANGELOG. Take the opportunity to use ListResource instead of
GetResource for some things that are closer to list resources than
individual resources.

Kevin took a look at the default clients and wasn't super happy with
how they handled this API, either.